Move key event and semantics related method from WidgetTester to WidgetController (#62362)
